+++ src/ng-vcl/checkbox/checkbox.component.ts
@@ -4,7 +4,6 @@ import { Component,
   OnChanges, SimpleChanges, EventEmitter,
   ChangeDetectionStrategy, ChangeDetectorRef, forwardRef, ElementRef
 } from '@angular/core';
-import { trigger } from '@angular/animations';
 import { Observable } from 'rxjs/Observable';
 import { ControlValueAccessor, NgControl, NG_VALUE_ACCESSOR } from '@angular/forms';
 
@@ -15,20 +14,18 @@ export const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR: any = {
 };
 
 @Component({
-  selector: 'vcl-checkbox',
+  selector: 'vcl-checkbox, [vcl-checkbox]',
   templateUrl: 'checkbox.component.html',
-  animations: [trigger('checkState', [])],
-  host: {
-    '[attr.role]': '"checkbox"',
-    '[class.vclCheckbox]': 'true',
-    '[style.userSelect]': '"none"'
-  },
   providers: [CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  host: {
+    '[class.vclInputInlineControlGroup]': 'inline === true',
+    '[class.vclInputControlGroup]': 'inline === false'
+  }
 })
 export class CheckboxComponent implements ControlValueAccessor {
 
-  @HostBinding()
+  @Input()
   tabindex = 0;
 
   @Input()
@@ -37,19 +34,18 @@ export class CheckboxComponent implements ControlValueAccessor {
   @Input()
   uncheckedIcon: string = 'fa:square-o';
 
-  @HostBinding('attr.aria-disabled')
-  @HostBinding('class.vclDisabled')
+  @Input()
+  label?: string;
+
   @Input()
   disabled: boolean = false;
 
   @Input()
-  labelPosition: 'left' | 'top' | 'right' = 'right';
+  inline: boolean = true;
 
-  /**
-  Reflects the checked state, `true` is checked and `false` is unchecked
-  @public
-  */
-  @HostBinding('attr.checked')
+  @Input()
+  iconPosition: 'left' | 'right' = 'left';
+
   @Input()
   checked: boolean = false;
 
@@ -59,9 +55,8 @@ export class CheckboxComponent implements ControlValueAccessor {
   @Output()
   checkedChange = new EventEmitter<boolean>();
 
-  constructor(private elementRef: ElementRef, private cdRef: ChangeDetectorRef) { }
+  constructor(private cdRef: ChangeDetectorRef) { }
 
-  @HostListener('keydown', ['$event'])
   onKeyup(e) {
     switch (e.code) {
       case 'Space':
@@ -78,7 +73,7 @@ export class CheckboxComponent implements ControlValueAccessor {
   }
 
   updateValue() {
-    if (this.disabled) {
+    if (this.isDisabled) {
       return;
     }
     this.checked = !this.checked;
@@ -91,7 +86,10 @@ export class CheckboxComponent implements ControlValueAccessor {
     return this.checked ? this.checkedIcon : this.uncheckedIcon;
   }
 
-  @HostListener('blur')
+  get isDisabled() {
+    return this.cvaDisabled || this.disabled;
+  }
+
   onBlur(e) {
     this.onTouched();
   }
@@ -112,8 +110,10 @@ export class CheckboxComponent implements ControlValueAccessor {
   registerOnTouched(fn: any) {
     this.onTouched = fn;
   }
+  // Store cva disabled state in an extra property to remember the old state after the radio group has been disabled
+  private cvaDisabled = false;
   setDisabledState(isDisabled: boolean) {
-    this.disabled = isDisabled;
+    this.cvaDisabled = isDisabled;
     this.cdRef.markForCheck();
   }
 }