+++ src/ng-vcl/checkbox/index.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { L10nModule } from './../l10n/index';
 import { VCLIconModule } from './../icon/index';
+import { VCLFormControlLabelModule } from './../form-control-label/index';
 import { CheckboxComponent } from './checkbox.component';
 
 export { CheckboxComponent };
 
 @NgModule({
-  imports: [CommonModule, VCLIconModule],
+  imports: [CommonModule, VCLIconModule, VCLFormControlLabelModule, L10nModule],
   exports: [CheckboxComponent],
   declarations: [CheckboxComponent]
 })