+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### Changes
 - (BREAKING) requires Angular ^5
+- checkbox:
+  - (BREAKING) Changed html structure
+  - (BREAKING) Removed support for labelPosition and changed to iconPosition
 
 ## 0.3.x (?)